cleaning old files; moved primary class definitions to nwb-conversion-tools
