+++ buzsaki_lab_to_nwb/original-code/BuzsakiLabNWBConverter.py
@@ -1,8 +1,7 @@
 
-from NWBConverter import NWBConverter
-from BaseDataInterface import NeuroscopeRecordingInterface
+from nwb_conversion_tools import NWBConverter, NeuroscopeDataInterface
 
 class BuzsakiLabNWBConverter(NWBConverter):
-    data_interface_classes = {'neuroscope': NeuroscopeRecordingInterface}
+    data_interface_classes = {'neuroscope': NeuroscopeDataInterface.NeuroscopeRecordingInterface}
