+++ buzsaki_lab_to_nwb/original-code/old_run_convert_yuta.py
@@ -1,6 +1,6 @@
 
 
-import convert_yuta
+import old_convert_yuta as convert_yuta
 
 convert_yuta.yuta2nwb(session_path='D:/BuzsakiData/SenzaiY/YutaMouse41/YutaMouse41-150903',
                                       include_spike_waveforms=True, stub=True)
